connect to database & update ui

+++ ConfigurationAnalyzer.Domain/Models/Basic/Resource.cs
@@ -2,10 +2,10 @@
 {
 	public class Resource
 	{
-		public long Id { get; set; }
+		public int Id { get; set; }
 
 		public string Name { get; set; }
 
-		public double Cost { get; set; }
+		public decimal Cost { get; set; }
 	}
 }
