+++ ConfigurationAnalyzer.Domain/Models/Run/ConfigurationRunProperties.cs
@@ -4,15 +4,15 @@ namespace ConfigurationAnalyzer.Domain.Models
 {
 	public class ConfigurationRunProperties
 	{
-		public long Id { get; set; }
+		public int Id { get; set; }
 
 		public Configuration Configuration { get; set; }
 
-		public double Duration { get; set; }
+		public int Duration { get; set; }
 
-		public double Cost { get; set; }
+		public decimal Cost { get; set; }
 
-		public double InefficiencyTime { get; set; }
+		public int InefficiencyTime { get; set; }
 
 		public IEnumerable<ProcedureRunProperties> Procedures { get; set; }
